dealt with seg fault error
