+++ string.c
@@ -31,7 +31,7 @@ char *_strcpy(char *dest, char *src)
 	{
 		dest[i] = src[i];
 	}
-	dest[i++] = '\0';
+	dest[i] = '\0';
 	return (dest);
 }
